+++ lib/fanoos_project.dart
@@ -1,2 +1,5 @@
 library fanoos_project;
 
+export 'src/app_service.dart';
+export 'src/service_container.dart';
+export 'src/widget_services.dart';