+++ lib/src/app_service.dart
@@ -1,6 +1,6 @@
 import 'package:scoped_model/scoped_model.dart';
 
-abstract class Service extends Model {
+abstract class AppService extends Model {
 
   Future<void> init(){}