+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [0.0.2]
+
+* exporting models
+* changed `Service` to `AppService`
+
 ## [0.0.1]
 
 * initial release.